Pose graph optimization now working, as well as mapping. TODO: publish the transform between the map and odom frames

+++ spinning_lidar_icpslam/include/utils/geometric_utils.h
@@ -76,9 +76,9 @@ tf::Transform getTFTransformFromPositionQuaternion(Eigen::Vector3d pos, Eigen::Q
 
 tf::Pose getTFPoseFromPositionQuaternion(Eigen::Vector3d pos, Eigen::Quaterniond q);
 
-Eigen::Vector3d getTranslationFromTMatrix(Eigen::Matrix4f &T);
+Eigen::Vector3d getTranslationFromTMatrix(Eigen::Matrix4d &T);
 
-Eigen::Quaterniond getQuaternionFromTMatrix(Eigen::Matrix4f &T);
+Eigen::Quaterniond getQuaternionFromTMatrix(Eigen::Matrix4d &T);
 
 geometry_msgs::Pose getROSPoseFromPosQuat(Eigen::Vector3d pos, Eigen::Quaterniond q);
